move all middleware into middleware.go
