+++ internal/api/middleware.go
@@ -1,0 +1,171 @@
+package api
+
+import (
+	"context"
+	"fmt"
+	"github.com/gin-gonic/gin"
+	"github.com/mcpjungle/mcpjungle/internal/model"
+	"github.com/mcpjungle/mcpjungle/internal/service/config"
+	"github.com/mcpjungle/mcpjungle/internal/service/mcp_client"
+	"github.com/mcpjungle/mcpjungle/internal/service/user"
+	"github.com/mcpjungle/mcpjungle/pkg/types"
+	"net/http"
+	"strings"
+)
+
+// requireInitialized is middleware to reject requests to certain routes if the server is not initialized
+func requireInitialized(configService *config.ServerConfigService) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		cfg, err := configService.GetConfig()
+		if err != nil || !cfg.Initialized {
+			c.AbortWithStatusJSON(http.StatusForbidden, gin.H{"error": "server not initialized"})
+			return
+		}
+		c.Next()
+	}
+}
+
+// verifyUserAuth is middleware that checks for a valid user token if the server is in production mode.
+// this middleware doesn't care about the role of the user, it just verifies that they're authenticated.
+func verifyUserAuthForAPIAccess(configService *config.ServerConfigService, userService *user.UserService) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		cfg, err := configService.GetConfig()
+		if err != nil {
+			c.AbortWithStatusJSON(
+				http.StatusServiceUnavailable, gin.H{"error": "failed to fetch server config while checking auth"},
+			)
+			return
+		}
+
+		c.Set("mode", cfg.Mode)
+
+		// no auth is required in case of dev mode
+		if cfg.Mode == model.ModeDev {
+			c.Next()
+			return
+		}
+
+		authHeader := c.GetHeader("Authorization")
+		token := strings.TrimPrefix(authHeader, "Bearer ")
+		if token == "" {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "missing access token"})
+			return
+		}
+
+		// Verify that the token is valid and corresponds to a user
+		authenticatedUser, err := userService.GetUserByAccessToken(token)
+		if err != nil {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "invalid access token"})
+			return
+		}
+
+		// Store user in context for potential role checks in subsequent handlers
+		c.Set("user", authenticatedUser)
+		c.Next()
+	}
+}
+
+// requireAdminUser is middleware that ensures the authenticated user has an admin role when in production mode.
+// It assumes that verifyUserAuthForAPIAccess middleware has already run and set the user in context.
+func requireAdminUser() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		mode, exists := c.Get("mode")
+		if !exists {
+			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": "server mode not found in context"})
+			return
+		}
+
+		m, ok := mode.(model.ServerMode)
+		if !ok {
+			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": "invalid server mode in context"})
+			return
+		}
+		if m == model.ModeDev {
+			// no admin check is required in dev mode
+			c.Next()
+			return
+		}
+
+		authenticatedUser, exists := c.Get("user")
+		if !exists {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "user is not authenticated"})
+			return
+		}
+
+		u, ok := authenticatedUser.(model.User)
+		if ok && u.Role == types.UserRoleAdmin {
+			c.Next()
+			return
+		}
+
+		c.AbortWithStatusJSON(http.StatusForbidden, gin.H{"error": "user is not authorized to perform this action"})
+	}
+}
+
+// checkAuthForMcpProxyAccess is middleware for MCP proxy that checks for a valid MCP client token
+// if the server is in production mode.
+// In development mode, mcp clients do not require auth to access the MCP proxy.
+func checkAuthForMcpProxyAccess(
+	configService *config.ServerConfigService,
+	mcpClientService *mcp_client.McpClientService,
+) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		cfg, err := configService.GetConfig()
+		if err != nil {
+			c.AbortWithStatusJSON(
+				http.StatusServiceUnavailable, gin.H{"error": "failed to fetch server config while checking mcp auth"},
+			)
+			return
+		}
+
+		// the gin context doesn't get passed down to the MCP proxy server, so we need to
+		// set values in the underlying request's context to be able to access them from proxy.
+		ctx := context.WithValue(c.Request.Context(), "mode", cfg.Mode)
+		c.Request = c.Request.WithContext(ctx)
+
+		if cfg.Mode == model.ModeDev {
+			c.Next()
+			return
+		}
+
+		authHeader := c.GetHeader("Authorization")
+		token := strings.TrimPrefix(authHeader, "Bearer ")
+		if token == "" {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "missing MCP client access token"})
+			return
+		}
+		client, err := mcpClientService.GetClientByToken(token)
+		if err != nil {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "invalid MCP client token"})
+			return
+		}
+
+		// inject the authenticated MCP client in context for the proxy to use
+		ctx = context.WithValue(c.Request.Context(), "client", client)
+		c.Request = c.Request.WithContext(ctx)
+
+		c.Next()
+	}
+}
+
+// requireServerMode is middleware that checks if the server is in a specific mode.
+// If not, the request is rejected with a 403 Forbidden status.
+func requireServerMode(configService *config.ServerConfigService, m model.ServerMode) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		cfg, err := configService.GetConfig()
+		if err != nil {
+			c.AbortWithStatusJSON(
+				http.StatusServiceUnavailable, gin.H{"error": "failed to fetch server config while checking mode"},
+			)
+			return
+		}
+		if cfg.Mode != m {
+			c.AbortWithStatusJSON(
+				http.StatusForbidden,
+				gin.H{"error": fmt.Sprintf("this request is only allowed in %s mode", m)},
+			)
+			return
+		}
+		c.Next()
+	}
+}
